requirements_cpu.txt
# CPU 전용 환경용 requirements (Fallback)
# PyTorch CPU 버전

--index-url https://download.pytorch.org/whl/cpu

# PyTorch CPU-only version
torch==2.5.0+cpu
torchvision==0.20.0+cpu
torchaudio==2.5.0+cpu

# Core ML and Deep Learning
timm==1.0.12
transformers==4.44.2
huggingface-hub==0.26.2

# Computer Vision
opencv-python==4.10.0.84
Pillow==10.4.0
albumentations==1.4.18
imageio==2.35.1

# OCR Support
easyocr==1.7.1
pytesseract==0.3.13
transformers[ocr]==4.44.2

# Data Science
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2

# Configuration and Utilities
pyyaml==6.0.2
tqdm==4.66.5
psutil==6.0.0
wandb==0.18.3

# Hyperparameter Optimization
optuna==4.0.0
ray[tune]==2.38.0

# Development Tools
jupyter==1.1.1
notebook==7.2.2
ipykernel==6.29.5

# Testing and Validation
pytest==8.3.3
pytest-cov==5.0.0

# CPU optimizations
accelerate==1.0.1
